Fixed the setup script to patch correctly.
[Details]
The setup script previously just git applied new changes for
defects4j_multi on top of any existing changes to the defects4j repo.
This however could conflict if the changes are incompatible. To fix
this, the setup script was adapted to remove all current changes to the
defects4j repo for the multi extension before re-applying them.

[File's changed]
	modified:   setup.py